Put grouping function in App
