+++ src/components/FiveDayForecast.tsx
@@ -7,17 +7,15 @@ import { Card } from './Card';
 import { Toggle } from './Toggle';
 import { WeatherIcon } from './WeatherIcon';
 import { round, degreeDescription } from '../utilities';
-import { Forecast, ForecastList, GroupedForecastList } from '../types/weather';
+import { ForecastList, GroupedForecastList } from '../types/weather';
 
 interface Props {
-  data: Forecast;
+  data: GroupedForecastList;
   unit: string;
 }
 
 export const FiveDayForecast: React.FC<Props> = ({ data, unit }: Props) => {
   const unitText = `°${unit === 'C' ? 'C' : 'F'}`;
-  let groupedLists: GroupedForecastList = {};
-
   const defaults = {
     color: 'text-gray-600',
     backgroundColor: 'bg-gray-800',
@@ -27,229 +25,216 @@ export const FiveDayForecast: React.FC<Props> = ({ data, unit }: Props) => {
       { background: 'linear-gradient(0deg, rgb(74, 85, 104) 90%, rgba(45,55,72,1) 100%)' },
   };
 
-  if (data.list) {
-    // Group lists by date
-    groupedLists = data.list.reduce((newItems: GroupedForecastList, item: ForecastList) => {
-      // Convert unix time string to year_month_date
-      const t = moment.unix(item.dt).format('YYYY_M_D');
-      if (!newItems.list) newItems.list = {};
-      if (!newItems.list[t]) newItems.list[t] = [];
-      newItems.list[t].push(item);
-
-      return newItems;
-    }, {});
+  if (!data.list) {
+    return <></>;
   }
 
   return (
     <>
-      {groupedLists.list ? (
-        <Card title="Forecast" titleSmall>
-          <ul className="grid grid-cols-1 gap-4">
-            {Object.entries(groupedLists.list).map(([key, day]: [string, ForecastList[]]) => {
-              // Highest temperature of the day
-              const highestTemp = Math.max.apply(
-                Math,
-                day.map((dayItem: ForecastList) => dayItem.main.temp),
-              );
+      <Card title="Forecast" titleSmall>
+        <ul className="grid grid-cols-1 gap-4">
+          {Object.entries(data.list).map(([key, day]: [string, ForecastList[]]) => {
+            // Highest temperature of the day
+            const highestTemp = Math.max.apply(
+              Math,
+              day.map((dayItem: ForecastList) => dayItem.main.temp),
+            );
               // Lowest temperature of the day
-              const lowestTemp = Math.min.apply(
-                Math,
-                day.map((dayItem: ForecastList) => dayItem.main.temp),
-              );
+            const lowestTemp = Math.min.apply(
+              Math,
+              day.map((dayItem: ForecastList) => dayItem.main.temp),
+            );
 
-              // Get mean values
-              const dayMean = {
-                temp: _.meanBy(day, (d: ForecastList) => d.main.temp),
-                humidity: _.meanBy(day, (d: ForecastList) => d.main.humidity),
-                pressure: _.meanBy(day, (d: ForecastList) => d.main.pressure),
-                speed: _.meanBy(day, (d: ForecastList) => d.wind.speed),
-              };
+            // Get mean values
+            const dayMean = {
+              temp: _.meanBy(day, (d: ForecastList) => d.main.temp),
+              humidity: _.meanBy(day, (d: ForecastList) => d.main.humidity),
+              pressure: _.meanBy(day, (d: ForecastList) => d.main.pressure),
+              speed: _.meanBy(day, (d: ForecastList) => d.wind.speed),
+            };
 
-              /*
+            /*
               * Find closest average values from mean value
               */
-              const chain = _.chain(day);
+            const chain = _.chain(day);
 
-              const dayAvgPressure = chain
-                .map('main')
-                .flatten()
-                .reduce((a, b) => (
-                  Math.abs(a.pressure - dayMean.pressure)
+            const dayAvgPressure = chain
+              .map('main')
+              .flatten()
+              .reduce((a, b) => (
+                Math.abs(a.pressure - dayMean.pressure)
                   <= Math.abs(b.pressure - dayMean.pressure)
-                    ? a : b
-                ))
-                .value();
-              const dayAvgHumidity = chain
-                .map('main')
-                .flatten()
-                .reduce((a, b) => (
-                  Math.abs(a.humidity - dayMean.humidity)
+                  ? a : b
+              ))
+              .value();
+            const dayAvgHumidity = chain
+              .map('main')
+              .flatten()
+              .reduce((a, b) => (
+                Math.abs(a.humidity - dayMean.humidity)
                   <= Math.abs(b.humidity - dayMean.humidity)
-                    ? a : b
-                ))
-                .value();
-              const dayAvgWind = chain
-                .map('wind')
-                .flatten()
-                .reduce((a, b) => (
-                  Math.abs(a.speed - dayMean.speed)
+                  ? a : b
+              ))
+              .value();
+            const dayAvgWind = chain
+              .map('wind')
+              .flatten()
+              .reduce((a, b) => (
+                Math.abs(a.speed - dayMean.speed)
                   <= Math.abs(b.speed - dayMean.speed)
-                    ? a : b
-                ))
-                .value();
-              const dayAvgTempIcon = chain
-                .reduce((a, b) => (
-                  Math.abs(a.main.temp - dayMean.temp)
+                  ? a : b
+              ))
+              .value();
+            const dayAvgTempIcon = chain
+              .reduce((a, b) => (
+                Math.abs(a.main.temp - dayMean.temp)
                   <= Math.abs(b.main.temp - dayMean.temp)
-                    ? a : b
-                ))
-                .value()
-                .weather[0].icon;
+                  ? a : b
+              ))
+              .value()
+              .weather[0].icon;
 
-              return (
-                <li key={key}>
-                  <Toggle
-                    content={(
-                      <div className={`flex flex-row flex-wrap items-center ${defaults.color} ${defaults.backgroundColor} py-2 px-4 rounded text-sm sm:text-base`}>
-                        <div className="flex-1">
-                          <p className="text-md text-white font-bold uppercase">
-                            {moment(key, 'YYYY_M_D').format('dddd')}
+            return (
+              <li key={key}>
+                <Toggle
+                  content={(
+                    <div className={`flex flex-row flex-wrap items-center ${defaults.color} ${defaults.backgroundColor} py-2 px-4 rounded text-sm sm:text-base`}>
+                      <div className="flex-1">
+                        <p className="text-md text-white font-bold uppercase">
+                          {moment(key, 'YYYY_M_D').format('dddd')}
+                        </p>
+                        <p className="text-sm text-gray-500">
+                          {moment(key, 'YYYY_M_D').format('MMMM Do')}
+                        </p>
+                      </div>
+                      <div className="flex-1">
+                        <WeatherIcon
+                          icon={dayAvgTempIcon}
+                          className="mx-auto"
+                        />
+                      </div>
+                      <div className="flex-1 flex flex-col flex-wrap items-center">
+                        <div className="w-12 whitespace-no-wrap text-right">
+                          <p>
+                            <span className="mr-1 text-gray-300 font-bold">
+                              {round(highestTemp)}
+                            </span>
+                            <span>{unitText}</span>
                           </p>
-                          <p className="text-sm text-gray-500">
-                            {moment(key, 'YYYY_M_D').format('MMMM Do')}
+                          <p>
+                            <span className="mr-1 text-gray-300">
+                              {round(lowestTemp)}
+                            </span>
+                            <span>{unitText}</span>
                           </p>
                         </div>
+                      </div>
+                      <div
+                        className="flex-1 text-center self-start"
+                        title="Average wind speed"
+                      >
+                        <FaWind className="mx-auto w-5 h-8" />
+                        <span className="text-gray-400 sm:whitespace-no-wrap text-sm">
+                          {`${round(
+                            dayAvgWind.speed,
+                          )} ${degreeDescription(dayAvgWind.deg)} `}
+                        </span>
+                        <span className="text-gray-600">
+                          {unit === 'C' ? 'm/s' : 'm/h'}
+                        </span>
+                      </div>
+                      <div
+                        className="flex-1 text-center self-start"
+                        title="Average humidity"
+                      >
+                        <WiHumidity className="mx-auto w-8 h-8" />
+                        <span className="text-gray-400 text-sm">
+                          {`${dayAvgHumidity.humidity} `}
+                          <span className="text-gray-600">%</span>
+                        </span>
+                      </div>
+                      <div
+                        className="flex-1 text-center self-start"
+                        title="Average pressure"
+                      >
+                        <WiBarometer className="mx-auto w-8 h-8" />
+                        <span className="text-gray-400 text-sm">{`${dayAvgPressure.pressure} `}</span>
+                        <span className="text-gray-600">hPa</span>
+                      </div>
+                    </div>
+                    )}
+                  toggledcontent={(
+                    <div
+                      className={`flex flex-row flex-wrap items-center ${defaults.toggledColor} ${defaults.toggledBackgroundColor} py-2 px-4 rounded`}
+                      style={defaults.toggledStyle}
+                    >
+                      <div className="flex-1">
+                        <p className="text-md text-white font-bold uppercase">
+                          {moment(key, 'YYYY_M_D').format('dddd')}
+                        </p>
+                        <p className="text-sm text-gray-500">
+                          {moment(key, 'YYYY_M_D').format('MMMM Do')}
+                        </p>
+                      </div>
+                      <div className="flex-1" />
+                      <div className="flex-1 " />
+                      <FaWind className="flex-1 w-8 h-5" title="Wind speed" />
+                      <WiHumidity className="flex-1 w-8 h-8" title="Humidity" />
+                      <WiBarometer className="flex-1 w-8 h-8" title="Pressure" />
+                    </div>
+                    )}
+                >
+                  <ul>
+                    {day.map((item: ForecastList) => (
+                      <li
+                        key={item.dt}
+                        className="flex flex-row flex-wrap items-center text-center text-gray-400 bg-gray-800 odd:bg-gray-900 px-4 rounded"
+                      >
+                        <div className="flex-1 text-gray-300 font-bold">
+                          {moment.unix(item.dt).format('HH:mm')}
+                        </div>
+
                         <div className="flex-1">
                           <WeatherIcon
-                            icon={dayAvgTempIcon}
+                            icon={item.weather[0].icon}
                             className="mx-auto"
                           />
                         </div>
-                        <div className="flex-1 flex flex-col flex-wrap items-center">
-                          <div className="w-12 whitespace-no-wrap text-right">
-                            <p>
-                              <span className="mr-1 text-gray-300 font-bold">
-                                {round(highestTemp)}
-                              </span>
-                              <span>{unitText}</span>
-                            </p>
-                            <p>
-                              <span className="mr-1 text-gray-300">
-                                {round(lowestTemp)}
-                              </span>
-                              <span>{unitText}</span>
-                            </p>
+                        <div className="flex-1 flex flex-wrap justify-center">
+                          <p className="w-12 whitespace-no-wrap text-right">
+                            <span className="text-gray-300 font-bold">
+                              {round(item.main.temp)}
+                            </span>
+                            <span className="text-gray-600">{` ${unitText}`}</span>
+                          </p>
+                        </div>
+                        <div className="flex-1 sm:whitespace-no-wrap">
+                          <div>
+                            {`${round(item.wind.speed)} ${degreeDescription(
+                              item.wind.deg,
+                            )}`}
+                          </div>
+                          <div className="text-gray-600">
+                            {unit === 'C' ? 'm/s' : 'm/h'}
                           </div>
                         </div>
-                        <div
-                          className="flex-1 text-center self-start"
-                          title="Average wind speed"
-                        >
-                          <FaWind className="mx-auto w-5 h-8" />
-                          <span className="text-gray-400 sm:whitespace-no-wrap text-sm">
-                            {`${round(
-                              dayAvgWind.speed,
-                            )} ${degreeDescription(dayAvgWind.deg)} `}
-                          </span>
-                          <span className="text-gray-600">
-                            {unit === 'C' ? 'm/s' : 'm/h'}
-                          </span>
+                        <div className="flex-1">
+                          {`${item.main.humidity} `}
+                          <span className="text-gray-600">%</span>
                         </div>
-                        <div
-                          className="flex-1 text-center self-start"
-                          title="Average humidity"
-                        >
-                          <WiHumidity className="mx-auto w-8 h-8" />
-                          <span className="text-gray-400 text-sm">
-                            {`${dayAvgHumidity.humidity} `}
-                            <span className="text-gray-600">%</span>
-                          </span>
-                        </div>
-                        <div
-                          className="flex-1 text-center self-start"
-                          title="Average pressure"
-                        >
-                          <WiBarometer className="mx-auto w-8 h-8" />
-                          <span className="text-gray-400 text-sm">{`${dayAvgPressure.pressure} `}</span>
+                        <div className="flex-1">
+                          {`${item.main.pressure} `}
                           <span className="text-gray-600">hPa</span>
                         </div>
-                      </div>
-                    )}
-                    toggledcontent={(
-                      <div
-                        className={`flex flex-row flex-wrap items-center ${defaults.toggledColor} ${defaults.toggledBackgroundColor} py-2 px-4 rounded`}
-                        style={defaults.toggledStyle}
-                      >
-                        <div className="flex-1">
-                          <p className="text-md text-white font-bold uppercase">
-                            {moment(key, 'YYYY_M_D').format('dddd')}
-                          </p>
-                          <p className="text-sm text-gray-500">
-                            {moment(key, 'YYYY_M_D').format('MMMM Do')}
-                          </p>
-                        </div>
-                        <div className="flex-1" />
-                        <div className="flex-1 " />
-                        <FaWind className="flex-1 w-8 h-5" title="Wind speed" />
-                        <WiHumidity className="flex-1 w-8 h-8" title="Humidity" />
-                        <WiBarometer className="flex-1 w-8 h-8" title="Pressure" />
-                      </div>
-                    )}
-                  >
-                    <ul>
-                      {day.map((item: ForecastList) => (
-                        <li
-                          key={item.dt}
-                          className="flex flex-row flex-wrap items-center text-center text-gray-400 bg-gray-800 odd:bg-gray-900 px-4 rounded"
-                        >
-                          <div className="flex-1 text-gray-300 font-bold">
-                            {moment.unix(item.dt).format('HH:mm')}
-                          </div>
-
-                          <div className="flex-1">
-                            <WeatherIcon
-                              icon={item.weather[0].icon}
-                              className="mx-auto"
-                            />
-                          </div>
-                          <div className="flex-1 flex flex-wrap justify-center">
-                            <p className="w-12 whitespace-no-wrap text-right">
-                              <span className="text-gray-300 font-bold">
-                                {round(item.main.temp)}
-                              </span>
-                              <span className="text-gray-600">{` ${unitText}`}</span>
-                            </p>
-                          </div>
-                          <div className="flex-1 sm:whitespace-no-wrap">
-                            <div>
-                              {`${round(item.wind.speed)} ${degreeDescription(
-                                item.wind.deg,
-                              )}`}
-                            </div>
-                            <div className="text-gray-600">
-                              {unit === 'C' ? 'm/s' : 'm/h'}
-                            </div>
-                          </div>
-                          <div className="flex-1">
-                            {`${item.main.humidity} `}
-                            <span className="text-gray-600">%</span>
-                          </div>
-                          <div className="flex-1">
-                            {`${item.main.pressure} `}
-                            <span className="text-gray-600">hPa</span>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  </Toggle>
-                </li>
-              );
-            })}
-          </ul>
-        </Card>
-      ) : (
-        ''
-      )}
+                      </li>
+                    ))}
+                  </ul>
+                </Toggle>
+              </li>
+            );
+          })}
+        </ul>
+      </Card>
     </>
   );
 };
